requirements-dev.txt
-r requirements.txt
pytest
pytest-django
pytest-cov